Fixed microcode editor &  improved readability
When copying a built in program to the source pane, the obect code wiull automatically fill in.
The line numbers area to the right of the source pane are now configurable programatically from colors.cpp.
The line number area now responds to dark mode.
The line numbers of the object code pane now are right justified.