Implemented basic ouput functionality
Added basic ability to write output to terminal and batch output widgets.
The output is currently fixed to 0xFEED, but this can be fixed in future versions.
There is no working input functionality.

Importantly, this means that the I/O ports need to be 16 bits wide instead of 8 bits wide, otherwise only one of the two bytes will ever be written to the output.