Execution Optimizations, Improved Debugger Granularity
Profiled application extensively with VerySleep to figure out hot spots in code.

This lead to some minor optimizations in CPUControlSection::branchHandler() & ::onRun(), ::onStep().

CPUDataSection was found to be the single slowest part of the program, so several improvements were added.

CPUDataSection ::setSignalsFromMicrocode was by far the slowest function, and its signalling functionality was never used.
So, this signaling on change functionality was removed for microcode controls & clocks.
For other signals, the can be enabled or disabled on the fly by calling setEmitEvents(bool).
Calculating ALU output was also incredibly slow, and it is done multiple time per step.
Caching the results of the ALU output calculation lead to a minor increase in performance.

CPUMemoizer now supports multiple logging levels. This allow greater debugging functionality when needed, or extra performance when it is not needed.

MainWindow now rembers if dark modw was enabled when launching.
MainWindow forces exclusivity of debugging levels under the statistics tab via a QActionGroup.

Added several compiler flags to Windows project configuration to generate PDB data needed for profiling.
Also tuned compiler flags for ~15% speed increase in release.
Switched to C++17 standard.