Microcode Breakpoints
Adjusted Line Area colors to be more visually distinct when used with breakpoints.

CPUControlSection is now able to hit breakpoints & signal to other components that it has hit a breakpoint.
For it to hit a breakpoint, the simulation must have been started with a call to onDebuggingStarted().
If onSimulationStarted is called, the program will execute normally, and will entirely ignore breakpoints.
When a breakpoint is hit, simulationHitMicroBreakpoint will be emitted.
When the UI has activated to handle the breakpoint, it must call onBreakpointHandled() so that debugging may continue.
Added lineHasBreakpoint() so that other modules may check if the CPU is breaking.

Fixed a rendering bug in the AMux & CMux tri-state labels.

Adjusted MainWindow to handle breakpoint events & re-rendering.

MemoryDumpPane now highlights correct PC, regardless of what the current µPC is.

Added breakpoints to MicroCode, cleaned up MicroCode class.

Formatted MicrocodePane's source code.
Added support for changing breakpoints on the fly via MicrocodePane.

Refactored MicroCodePane to follow style conventions.