Fixed bugs in microcode
There were 3 major bugs in the microcode that needed to be addressed.

The implementation for the MOVSPA instruction reference the wrong register as its source (1-accumulator lo), and this has been corrected to be (4-SP hi).

Fixed a bug in the negx instruction where the andz flag was set so that the z bit would be true iff the high order byte was zero.

Fixed a bug in the CPWA, CPWX instructions. According to the book, if there is a signed overflow, the instructions must negate the N bit. this is handled by the block at the symbol cpw_n_fix. If the V bit is not true, then there is no need to perform this correction, and execution continues normally.