Added text highlighting to Printer
Created an HTMLHighlighterMixin, which allows a Qt syntax highlighter to generate stylized text from a plain text document.
The main window was modified to use html generating methods for printing. This allows stylized text to be printed, instead of the default black text.
Microcode  Pane now keeps track if it is in dark mode, and can now generate stylized text for printing.
ASM and microcode highlighter now derive from HTMLHighlighterMixin, so as to make generating stylized text for printing easier.