Fixed linFIb example bug
Fixed a bug in the linFib program where generating numbers larger than 0x4000 would cause the program to freeze.
This was caused by an erronious right shift causing unintended sign extension.
This is fixed by using an extra load in lieu of a shift.
Comments are updated to reflet changed code.