ROM module now burned in correctly
Modified assembler to provide additional information. Operating system is now aware of the location at which it must be burned.
Object code is no longer generated it it occurs before a burn.

Application now longer crashes when writing to read only memory.

Application should display popup when a simulation finished with an error.