Microcode bug fix, redefine mnemonics bugfixes
Added additional mnemonic map in pep.h to prevent microcode from being effected by the redefinition of mnemonics.
Renaming NOP1 to NOTY previously made microcode error when the using the ISD decoder on instruction 0x27.
This behavior is now fixed.

Fixed a bug where the microcode for the trap instruction would improperly flag NOP1 as a non-unary no-op.

The redefine mnemonics tab caused issues with the implementation of the ISD in the CPU.
The ISD looked up the string name of the instruction in register 8.
Redefined mnemonics had differing names than what was expected in the microcode, causing the ISD to fail unexpectedly.
This is fixed with a jump table that is calculated at the beggining of CPU execution, instead of being computerd dynamically.