Reduce number of memory chip allocations
Number of memory chips allocated over execution of program is vastly reduced.

This was achieved by creating a helper function that took a list of chip locations & sizes.
Using this list, it re-uses chips already installed in memory, and only discards chips that will not be used over a run.

This required making all memory chips resizable.