Silence compile warnings, fix IO bugs
Fix multiple bugs regarding IO. WIth the refactoring of the IO ports, the IOWIdget was no longer sucessfully aborting IO.
The IOWidget also wrote the contents of the batch input to memory twice, which was problematic.
MainWIndow wasn't loading OS before assembling programs, which caused bad values for charIn, charOut to be generated.
MainMemory now caches which addresses have pending IO, so that they may be canceled.