Improve modularity in simulator
Originally, the microcode simulator assumed that the address of 0 was where the von neuman cycle started.
This is incorrect, as certain sequences of instructions may need to be executed before a program.
For example, loading the stack pointer needs to be done in hardware, not the sim.
For now, the simulator still uses the "cheat" of setting the upc, but this is a step in the right direction.

Add ability to fetch from memory vectors from program manager, as calculating the offsets manually was error prone.