Create Pep9, fix suite wide crash
MainMemory is a child of the MainWindow in all Pep applications. This is incorrect, as it is owned by a smart pointer.
In creating Pep9, the application crashed because the object was deleted twice.
Fixed this bug in all Pep applications.

Pep9 has been cloned from the Pep9Micro sources.