Fix invalid microcode being executed
In Pep9Micro/CPU, there is a bug where invalid microcode could be executed.
This would happen after the following steps:
1) A valid microprogram was built.
2) The file was saved.
3) An invalid microprogram was written.
4) The file was saved.
5) The simulator was started.
To fix this, the micrprogram cache is cleared every time a file is saved.
As the microprograms are lengthy, it is still beneficial to cache the microprogram in most cases.