Help Dialog no longer prevents application from closing
The help dialog spawns in a separate window. When the main application closed, it would not be able to shut down entirely until after the help dialog closed.
To fix this, when a close event is received, it needs to kill the help dialog.
All 3 applications must do this.
None of the other pop-ups have this behavior, so this fix is only needed for the help dialog.