Fix CPWr, newline parsing bugs.
Fixed a bug in Pep9Term where object code bytes separated by \n wouldn't be interpreted as valid bytes.

Fixed a bug in the implementation of CPWr, CPBr, owing to the representation of the operand.
In Pep9Micro & Pep9 (original), the 2's complement of the operand  is used, where as Pep9asm uses the original operand and subtracts it.
This leads to the NV bits being incorrect a significant portion of times, and is now fixed.
Added Pep9term to pep suite build file.