Fix file output race condition.
Fixed a race condition that occurred in Pep9term with IO.
When a simulation would finish, it could quit before pending IO had finished.
This would cause pending IO to fault, and caused certain objects to fault from bad memory access.
This seems to be fixed via keeping pointer to the file being written to (this allows us to block on write),
and opting to not close the file until the file object is deleted.