Allow smaller burn values in Pep9.
Smaller values than 0xFFFF are now allowed in Pep9asm.
In order to implement this, many changes had to be made.

First, it was discovered that lookup tables in main memory were doing a bad job of caching pointers.
This lead to lots of crashes when resizing memory, and has since been fixed by forcing pointer caches to update on any configuration changes.
Also, changed from passing around shared pointers to raw pointers, as there was an upcasting error being hidden by this lookup.
Changed size() => maxAddress(), as the only computationally useful value is the max accessible address.
The size can be calculated as max + 1.

Differentiated between crashes on NilChip, but will still crash for now. This is a helpful feature (hah) for debugging.
Same problem exists of writing to ROM, input chips.

With these changes, the memory dump pane can now respect the maximum size of memory, even if it is less than 0xFFFF.