Enable interrupt execution.
Interrupt execution was previously implemented as stopDebugging, which was incorrect behavior.
What is should have done is pause program execution & start debugging the running program.
To achieve this, several changes were needed across the project.

First, the CPU model needed the ability to manually “force” a breakpoint to occur.
This required rethinking how single steps & multi steps interacted with the CPU model.
Instead of checking for breakpoints in the run() function, the low-level atomic step function must perform this check.
This is the onMCStep in pep9Micro & onISAStep in Pep9.
By performing the check in the step function, any sequence of steps that requires multiple cycles will properly interact with the event loop.

In addition, the CPU needed the ability to enter debugging mode after having been started.
This required removing the old onStartDebugging() onFinishedDebugging() options, which reset all CPU state.
In their place, an enableDebugging() function was added, which simply tells the CPU to start listening to debug events.

With these changes, it is now possible to use the conditioned step functions to step until a breakpoint is hit.