Add copy functionality to memory dump pane.
Add the ability to copy text out of the memory dump pane.
This behavior existed in the old Pep9 memory dump pane, but was dropped when switching from a text edit to a table view.
The old functionality was useful, as it helped create text based diagrams for memory dumps.
So, created a copy functionality that is identical in behavior to Pep9 while utilizing the new components.