Properly decode operand value in Asm CPU pane.
Added additional caching features to Pep9 / micro for preserving OpSpec values.

If one were to re-decode the value of OpSpec at the end of a cycle for the UI, one might not get the intended value.
This could occur with STWA ,d. Modifying the contents of memory might change the decoded value of the operand.
To avoid this, just cache the decoded OpSpec during step.