Fix bug where printing in Pep9CPU crashes program.
This commit fixes issues #54 for Pep9CPU.
There was a bug where attempting to print microcode would crash the application.
This occurred because the Pep9CPU code highlighter used for printing assumed that the full control section of Pep9Micro was present.
Pep9CPU doesn’t support these features, and as such many of the enumeration maps were left initialized.
These initialized maps where then included in the highlighting rules - as the empty string.
The regular expression matcher matched the empty string to everything, cause the application to crash.

This fix to this bug is to disable Pep9Micro - specific highlighting.