Change “Format from listing”→“Assemble & Format”.
Modify the behavior of the format from listing options in Pep9/CPU/Micro.
Before, there had to exist an already assembled program for these features to work.
This was confusing, as formatting might not reflect the most recent changes to a program.

Now, whenever “assemble & format” is hit, the source code will unconditionally be recompiled.
Then the source code will be formatted correctly IF the source code assembled successfully.

This name change should make the UX of formatting a little more straightforward.