Fix bug allowing malformed trace tags to assemble.
If a line has a symbolic or format tag, we must perform additional parsing.
Failure to check for symbol tags led to bug #80, where programs only containing symbolic trace tags would assemble with no warnings.
So, if a symbolic trace tag is detected, we must note that trace tags are present in the program.
This allows handleTraceTags(...) to detect the cases where a symbolic tag is undefined.

Fixes bug #80.