UI consistency changes
I've tweaked spacing on the ASM UI so spacing is closer to what it was pre-pep9suite. I also added a new "stop debugging" icon, and updated the "step" icons to have better transparency. They're slightly larger in dimensions, so please let me know if they don't render correctly on Mac.

I also made it so the execution statistics pane shows if it has focus now. Lots of little UI tweaks all around.

These changes are largely for ASM and not CPU/Micro, as this is all I had time to get to tonight.