Implement UI for cache.
Implement a bare-bones UI for the cache.
This imitates the UI look-and-feel for the Memory Dump Pane.
This UI is currently the way to configure the cache, but this might change in future versions.

At this point, all the caching algorithms other the "Random" need to be debugged.
They seem to get strange results compared to hit/miss numbers.

Changes to Pep9Asm need to be backported to Pep9Micro, especially after the execution statistics widget is modified.
Sizing in the CacheView is weird, scaling needs to be adjusted.
The ISAMemoizer needs to become aware of the cache and clearing the read bytes list.

Tree PLRU, 1-bit PLRU, and ARC are all algorithms that should be implemented after debugging / commenting current code.
Additional features include concurrent execution of multiple caching algorithms to provide better statistics.
A heat map for memory access would also be nice.