Utilize lookahead in during cache replacement.
Prior to this commit, the next eviction item was calculated in multiple functions.
This increases the chance that the item evicted is not the actual target victim.
Changes to the AReplacmentPolicy class added a single lookahead function to get the next victim.
This can be re-used by evict() to avoid computign the next victim.

Streamlined CacheView interface.