Improve speed of memory dump pane.
Speed issues have been reported in the simulation.
During the creation of Pep9Cache, I discovered the issue was caused by refreshing the memory dump pane's column sizes too often.
Moving the resizeColumnsToContests(...) calls out of the hot path in refreshMemoryLines(uint16,uint16) fixed the bottleneck.
This will cause a merge conflict with Pep9Cache, but this fix is needed immediately as we plan on releasing Pep9 v9.3 soon.