Split FHS creation from app deployment on Linux.
Some systems (Ubuntu 16.04) don't make it easy to link system Qt libs for linuxdeployqt.
Rather than attempting to debug the linker failure, it is easier to download a working LinuxDeployQt appimage and invoke it myself.
To support this, creating a FHS filesystem needs to be split from app deployment, like repository generation was split from installer creation.
This feature is enabled with DEPLOY_OPT+="linux_fhs".